backend/requirements.txt
fastapi==0.128.0
sqlmodel==0.0.31
uvicorn==0.40.0
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.21
psycopg2-binary==2.9.11
pydantic==2.9.2
typing-extensions==4.15.0
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.25.0